Update patch set 1
Patch Set 1:

Probably not; I think we'll need different jobs for that case (they might work as-is with the plugin-build job, or we might need a third job to handle that case).

Depending on the urgency, we could land this as-is and then work to correct that situation while ignoring the expected failures.  Or we could alter this change to exclude those branches from those plugins.  Or we could reduce this change to only the plugins that are core in all 4 branches, then expand from there.

Patch-set: 1