Update patch set 1
Patch Set 1:

The current state is returning failing results on changes to download-commands.  Here's an example: https://ci.gerritcodereview.com/t/gerrit/build/753ed28956bb474da16d70351fba6d37

I agree with you, there's a danger in accidentially having the CI system not run tests when they exist.  One option would be to have the job automatically determine whether or not to run tests based on the contents of the repo.  However, when I've seen that done in the past, I have also seen projects go months without testing because that check silently failed.  For that reason, I recommend configuring the job explicitly for the tests we expect to run.

Another way to mitigate this is to put this configuration in the repo itself, so that in the same commit when tests are added, the CI configuration can be changed in order to enable them, and reviewing both the tests, the CI configuration, and the test results can all be done in the same change.  Even though this configuration is in a central repo, there's nothing to prevent us from doing that (even temporarily) -- an in-repo configuration will override this one.

Patch-set: 1