Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The current state is returning failing results on changes to download-commands.  Here's an example: https://ci.gerritcodereview.com/t/gerrit/build/753ed28956bb474da16d70351fba6d37
> 

Ah. This is because bazel returns a non-zero exit status when bazel test is run but there were no tests to be executed.

Can we change the job so that it detects the exit code?  It's a different code in this case than it is when the tests were run but failed.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>