Update patch set 1
Patch Set 1:

Replied on the other change (sorry I was in the middle of combing through logs); in short, it looks like a transient git fetch error.

I think we can make the checker required for submit, though we were also talking about having Zuul report Verified, since there are still some benefits to that (showing up in the dashboard view, and updating the last updated time for the change).  I was about to push up a change for that; if Zuul were reporting Verified, would you still want to drop it for code-owners?

Patch-set: 1