Revert "Remove plugins/task job as it defines its own job"
This reverts commit 52ce37fbe139fe62eb87a4bd3514162948413408.

Zuul doesn't seem to be picking up task plugin changes on any branch
except 2.16, which was the only branch that had a .zuul.yaml when the
reverted commit was merged. Hopefully re-adding a task plugin job will
force Zuul to notice the .zuul.yaml files that are now present on all
the task plugin branches.

Change-Id: I05c62d564c950ea3f030a646b2f383ba0cb93c20