Update the apt indices before installing packages
Images may sometimes ship with outdated package data, so installing
something in that state could lead to failures when a cached index
refers to a package version which is no longer available from the
repository. Force an update of the indices first in order to avoid
that potential failure condition.

Change-Id: I54bd13f77ea4e53d11efbc3208a878c331ea3291